pyproject.toml
[project]
name = "llm-wiki"
version = "0.1.0"
description = "A personal knowledge base that builds itself — LLM incrementally compiles knowledge into an interlinked wiki"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "litellm>=1.40",
    "rich>=13.7",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "python-frontmatter>=1.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
vector = [
    "lancedb>=0.6",
    "sentence-transformers>=3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
wiki = "llm_wiki.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
